extract BookPageConversations component
