+++ tailwind.config.js
@@ -63,6 +63,7 @@ module.exports = {
         700: "hsl(45, 8%, 37%)",
         800: "hsl(45, 8%, 29%)",
         900: "hsl(45, 8%, 22%)",
+        950: "hsl(45, 8%, 16%)",
       },
     },
     extend: {
